requirements.txt
hydra-core>=1.3.2
omegaconf>=2.3.0
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.3.0
torchmetrics>=1.4.0
transformers>=4.44.0
datasets>=2.19.0
accelerate>=0.34.0
peft>=0.12.0
hessian-eigenthings>=0.3.3
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.2.0
tensorboard>=2.16.0
scikit-learn>=1.4.0
pytest>=8.0.0
